Move jxbrowser setup to separate class
